# swaggerui/swagger.yml
openapi: 3.0.0
info:
  description: This is documentation to use Dorayaki backend API
  version: 1.0.0
  title: Swagger Dorayaki backend API
  contact:
    email: 13519105@std.stei.itb.ac.id
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /shops/{shopId}:
    get:
      tags:
        - Shops
      summary: Get shop/toko by ID
      description: Get shop/toko and its all dorayakis
      operationId: getShop
      parameters:
        - in: path
          name: shopId
          required: true
          description: ID of the shop
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Get shop/toko success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokoDataResponse'
        404:
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotFound'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
    put:
      tags:
       - Shops
      summary: Edit shop by id
      description: Edit base information about shop/toko (not including its dorayakis)
      operationId: updateShop
      security:
        - BearerAuth: []
          CookieAuth: []
          QueryAuth: []
      parameters:
        - in: path
          name: shopId
          required: true
          description: ID of the shop
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputToko'
              description: Input new shop's data
      responses:
        200:
          description: Update shop success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokoDataResponse'
        400:
          description: Error because invalid body request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        401:
          description: Error because wthere is no JWT token. Check if its exist and valid in order of Bearer, Cookie, and then Query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
    delete:
      tags:
       - Shops
      summary: Delete shop by id
      description: Delete shop/toko by id and its all dorayakis
      operationId: deleteShop
      security:
        - BearerAuth: []
          CookieAuth: []
          QueryAuth: []
      parameters:
        - in: path
          name: shopId
          required: true
          description: ID of the shop
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Delete success
        404:
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotFound'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /shops/{shopId}/stocks/{dorayakiId}:
    get:
      tags:
        - Shops
        - Stocks
      summary: Get stock of dorayaki in the shop
      description: Get stock of dorayaki in the shop. If the dorayaki was not added in the shop, then this endpoint will add that dorayaki in the shop with stock = 0.
      operationId: getStock
      parameters:
        - in: path
          name: shopId
          required: true
          description: ID of the shop
          schema:
            type: integer
            minimum: 1
        - in: path
          name: dorayakiId
          required: true
          description: ID of the dorayaki
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Get stock success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockDataResponse'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
    post:
      tags:
        - Shops
        - Stocks
      summary: Add/reduce stock of dorayaki in the shop
      description: Add stock of dorayaki in the shop. If want to reduce it, pass a negative value
      operationId: addStock
      security:
        - BearerAuth: []
          CookieAuth: []
          QueryAuth: []
      parameters:
        - in: path
          name: shopId
          required: true
          description: ID of the shop
          schema:
            type: integer
            minimum: 1
        - in: path
          name: dorayakiId
          required: true
          description: ID of the dorayaki
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                add_stock:
                  type: integer
                  description: Amount of dorayaki that want to be added. Pass a negative value if want to reduce it.
      responses:
        200:
          description: Add/reduce stock success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockDataResponse'
        400:
          description: Error because invalid body request (reducing stock to be <0 will trigger this error too)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        401:
          description: Error because wthere is no JWT token. Check if its exist and valid in order of Bearer, Cookie, and then Query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        500:
          description: Error because something wrong with internal server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
    QueryAuth:
      type: apiKey
      in: query
      name: jwt
  schemas:
    InputToko:
      description: Schema for input shop/toko
      type: object
      properties:
        nama:
          type: string
          description: Name of shop/toko
        jalan:
          type: string
          description: Street address of shop/toko
        kecamatan:
          type: string
          description: Subdistrict of shop/toko
        provinsi:
          type: string
          description: District/province of shop/toko
      required:
        - nama
        - jalan
        - kecamatan
        - provinsi
...
    Dorayaki:
      description: Schema for dorayaki
      type: object
      properties:
        id:
          type: integer
          description: ID of dorayaki
        rasa:
          type: string
          description: Taste of dorayaki
        deskripsi:
          type: string
          description: Description of dorayaki
        image_url:
          type: string
          description: Image URL of dorayaki
        created_at:
          type: integer
          description: Time since dorayaki's variant created
        updated_at:
          type: integer
          description: Time since dorayaki's variant updated
      required:
        - rasa
        - deskripsi
    Toko:
      description: Schema for shop/toko
      type: object
      properties:
        id:
          type: integer
          description: ID of shop/toko
        nama:
          type: string
          description: Name of shop/toko
        jalan:
          type: string
          description: Street address of shop/toko
        kecamatan:
          type: string
          description: Subdistrict of shop/toko
        provinsi:
          type: string
          description: District/province of shop/toko
        dorayaki:
          type: array
          items:
            $ref: '#/components/schemas/Dorayaki'
        stok:
          type: array
          items:
            $ref: '#/components/schemas/TokoDorayaki'
        created_at:
          type: integer
          description: Time since the shop created
        updated_at:
          type: integer
          description: Time since the shop updated
      required:
        - nama
        - jalan
        - kecamatan
        - provinsi
    TokoDorayaki:
      description: Schema for dorayaki's stock in the shop
      type: object
      properties:
        toko_id:
          type: integer
          description: ID of shop/toko
        dorayaki_id:
          type: integer
          description: ID of dorayaki
        stok:
          type: integer
          description: Stock of dorayaki in the shop
        created_at:
          type: integer
          description: Time since dorayaki's variant added to the shop
        updated_at:
          type: integer
          description: Time since dorayaki's variant restocked to the shop
    ErrorBadRequest:
      description: Bad request
      type: object
      properties:
        status_code:
          type: integer
          description: Error's status code
          default: 400
        message:
          type: string
          description: Error's message
          default: Bad request
    ErrorUnauthorized:
      description: There is no token in header, cookie, or url query
      type: object
      properties:
        status_code:
          type: integer
          description: Error's status code
          default: 401
        message:
          type: string
          description: Error's message
          default: Unauthorized
    ErrorNotFound:
      description: Resource not found
      type: object
      properties:
        status_code:
          type: integer
          description: Error's status code
          default: 404
        message:
          type: string
          description: Error's message
          default: Resource not found
    ErrorInternalServer:
      description: Internal server error
      type: object
      properties:
        status_code:
          type: integer
          description: Error's status code
          default: 500
        message:
          type: string
          description: Error's message
          default: Internal server error
...
    TokoDataResponse:
      description: Response that its data contains Toko/shop schema.
      type: object
      properties:
        status_code:
          type: integer
          description: Status code of response
          default: 200
        message:
          type: string
          description: Message of response
          default: Success
        data:
          type: array
          items:
            $ref: '#/components/schemas/Toko'
    StockDataResponse:
      description: Response that its data contains TokoDorayaki (stock) schema.
      type: object
      properties:
        status_code:
          type: integer
          description: Status code of response
          default: 200
        message:
          type: string
          description: Message of response
          default: Success
        data:
          type: array
          items:
            $ref: '#/components/schemas/TokoDorayaki'
...